==> code/nitze_regions/13_visualize_composites.ipynb <==
# Visualize Planet Composite Images

import ee
ee.Initialize()

# Import Libraries
import geemap
import re
from google.cloud import storage
from pprint import pprint

# Set up access to abrupt_thaw
storage_client = storage.Client(project="AbruptThawMapping")
abrupt_thaw = storage_client.get_bucket('abrupt_thaw')

## Define Functions

# apply image IDs to each image in the collection
def setID(image):
    img_id = image.id();
    img_prop = image.setMulti({'ID': img_id});
    img_prop = ee.Image(img_prop);
    return img_prop;

# Mask the data to values greater than 0 (which is how nodata gets imported into GEE by default)
def mask_0(image):
    mask = image.gt(0)
    return image.updateMask(mask)

## Import Data

# Import Planet Data
planet = ee.ImageCollection('projects/abruptthawmapping/assets/nitze_regions_composites')
planet = planet.map(setID)

# Import Sentinel Data
s2 = ee.ImageCollection('projects/abruptthawmapping/assets/nitze_regions_sentinel_reference')
s2 = s2.map(setID)

# Polygons
polygons = ee.FeatureCollection('users/gfiske/FrostCraters/rts_polygons_in_Nitze_areas_for_Yili_Mar_2022')

# Filter to look at composites which include both 2019 and 2018 images
planet_2019 = planet.filter(ee.Filter(ee.Filter.stringContains('ID', '2019'))).median()
planet_all = planet.filter(ee.Filter(ee.Filter.stringContains('ID', 'all'))).median()
# pprint(planet.aggregate_array('ID').getInfo())

# Prep Map
Map = geemap.Map()
Map.centerObject(planet)

# Set visualization parameters
vis_params = {
    'min': 0,'max': 1200,
    'bands': ['b3', 'b2', 'b1'],
    'gamma': 1
}

# Add layers
# Add Sentinel-2 composite
Map.addLayer(s2.median(), {'bands':['B4', 'B3', 'B2'], 'min':0, 'max':1200}, 'S2 All Years')

# add Planet composites
Map.addLayer(planet_2019, vis_params, 'Planet 2019')
Map.addLayer(planet_all, vis_params, 'Planet All')

# add RTS Polygons
Map.addLayer(polygons, {'color': 'red'}, 'RTS')

# Display the map
Map
